Bugfix: VRK data does not specify a unique id for OpenAddresses.
The id which comes with VRK based OpenAddresses data seems to specify
a building. A large building can open to 5 - 10 different street
addresses. Consiquently, a remarkable part of addresses were lost
when indexed to Elasticsearch using the non-unique id.

Now the id is a combo of building id and a hash from
zip, number and location coords.